+++ NEWS.md
@@ -1,5 +1,7 @@
 # gdalraster 1.5.1 (dev)
 
+* add `gdal_formats()`: report the supported raster formats (2023-09-28)
+
 * add `getCreationOptions()`: get the list of creation options of a raster format (2023-09-28)
 
 * add `copyDatasetFiles()`: copy all the files associated with a dataset (2023-09-28)